Fix full table recovery
During a full table sync, the records were being order by xmin _as
text_. This means that a row with xmin 10 will be processed before one
with xmin 2.

It's possible for 10, 11, etc. to be processed before 2. As an example,
it could process a row with xmin 12, update the bookmark, then fail.

During a full table sync recovery it would only look for rows >= 12
_numerically_, which means that a row with xmin 2 would never get
processed.